feat: 重構狀態管理與環境配置模組
- 將保存狀態相關的輔助函式移至共享模組，並從原有位置導出，提升代碼的可重用性與可維護性。
- 新增環境檢測配置的聚合入口，為未來的環境相關常量遷移做準備。
- 新增擴展專用的配置聚合入口，確保擴展頁面與背景使用的配置集中管理。
- 新增高亮樣式相關的常量模組，集中管理高亮顏色與匹配分數的邏輯。
- 新增 Notion 語言相關的共享常數，避免語言白名單與回退邏輯分散，提升一致性。
- 新增 UI 圖標配置模組，集中管理所有系統使用的 SVG 圖標字串，實現樣式與內容的分離。
