+++ scripts/config/extension/index.js
@@ -1,0 +1,12 @@
+/**
+ * Extension-only 配置聚合入口
+ *
+ * 此目錄僅供 extension pages / Background 使用，
+ * Content Script bundle 不應引入此入口。
+ */
+
+export * from './authMode.js';
+export * from './notionAuth.js';
+export * from './notionApi.js';
+export * from './accountApi.js';
+export * from './driveSyncErrorCodes.js';
