+++ scripts/config/shared/highlight/index.js
@@ -1,0 +1,1 @@
+export * from './style.js';
