+++ scripts/config/shared/index.js
@@ -1,0 +1,4 @@
+export * from './status/index.js';
+export * from './notion/index.js';
+export * from './highlight/index.js';
+export * from './ui/index.js';
