+++ scripts/config/shared/notion/index.js
@@ -1,0 +1,1 @@
+export * from './codeLanguages.js';
