+++ scripts/config/shared/ui/index.js
@@ -1,0 +1,1 @@
+export * from './icons.js';
